VIX-3625 Video Effect Cache Improvements
- Changed the cache folder model from InstanceID based to Settings Hash base with an associating file - In VideoEffect temp folder, (InstanceID).(Settings Hash)
This allows the same settings to build or use the same cache folder, so multiple effects can use the same cache folder too
Used MD5 to create the hash so it is consistent between executions of the application
- Added new GetVideoDurationAndResolution function that doesn't use the cache folder and gets both in a single run of ffmpeg
- Refactored Video.cs:GetVideoInformation to use GetVideoDurationAndResolution
- Added a Removing method to allow TimedSequenceEditorForm to let an Effect know when it has been deleted
- Updated Dispose method to clean up cache folder
- Used path functions in Video.cs
- Tweaked error message about Video Length
- Changed VideoLength from int to double
- Marked the unused ffmpeg public functions as Obsolete

+++ src/Vixen.Modules/Effect/Video/VideoData.cs
@@ -97,7 +97,7 @@ namespace VixenModules.Effect.Video
 		public Curve LevelCurve { get; set; }
 
 		[DataMember]
-		public int VideoLength { get; set; }
+		public double VideoLength { get; set; }
 
 		[DataMember] 
 		public int MovementRate { get; set; }
